Updated methods and spell check
GetKeyBinding method was searching the table for a username based on the value stored in the "command" variable instead of the value stored in the "username" variable. This led to the method returning an empty string instead of the value being stored in the table for the given command. I have changed the method to query the table based off of the correct value, and it is not returning the proper value.